reafactor(update): pangolin scripts were adapted for the new fee system
