+++ ignition/deployments/chain-43113/artifacts/KOLFactoryPangolinModule#KOLFactoryPangolin.dbg.json
@@ -1,0 +1,4 @@
+{
+  "_format": "hh-sol-dbg-1",
+  "buildInfo": "../build-info/b38c63001ba72232d8601cf239fb2432.json"
+}
