+++ package.json
@@ -4,7 +4,6 @@
     "deploy:whreceiver": "npx hardhat ignition deploy ignition/modules/wormhole/SL1MessageReceiver.ts",
     "deploy:whsender": "npx hardhat ignition deploy ignition/modules/wormhole/SL1MessageSender.ts",
     "deploy:koltraderjoe": "npx hardhat ignition deploy ignition/modules/KOLFactoryTraderJoe.ts",
-    "deploy:koluniswap": "npx hardhat ignition deploy ignition/modules/KOLFactoryUniswap.ts",
     "deploy:kolpangolin": "npx hardhat ignition deploy ignition/modules/KOLFactoryPangolin.ts",
     "deploy:kolarenaswap": "npx hardhat ignition deploy ignition/modules/KOLFactoryArenaSwap.ts",
     "register:whreceiver": "npx hardhat run scripts/wormhole/RegisterSenderInReceiver.ts",
@@ -44,11 +43,6 @@
     "@pangolindex/sdk": "^5.8.4",
     "@traderjoe-xyz/sdk-core": "^2.0.11",
     "@traderjoe-xyz/sdk-v2": "^3.0.26",
-    "@uniswap/permit2-sdk": "^1.3.1",
-    "@uniswap/sdk-core": "^7.7.2",
-    "@uniswap/smart-order-router": "^4.21.2",
-    "@uniswap/universal-router-sdk": "^4.19.5",
-    "@uniswap/v3-sdk": "^3.25.2",
     "dotenv": "^16.4.5",
     "fs": "^0.0.1-security",
     "ts-node": "^10.9.2",
