+++ scripts/kol-router/pangolin/createKolRouter.ts
@@ -6,7 +6,6 @@ import { avalancheFuji } from 'viem/chains';
 async function main() {
     const privateKey = process.env.DEPLOYER_KEY as string;
     const account = privateKeyToAccount(`0x${privateKey.slice(2)}`);
-    const fee = "30000000000000";
     // Initialize clients
     // ---------------------------------------------------------------------
     const walletClient = createWalletClient({
@@ -15,7 +14,7 @@ async function main() {
         transport: http(),
     });
 
-    await createKolRouter(account.address, fee, walletClient);
+    await createKolRouter(account.address, walletClient);
 }
 
 main().catch((error) => {
