--- scripts/kol-router/uniswap/swapExactNATIVEForToken.ts
@@ -1,23 +1,0 @@
-import { executeSwap } from './functions';
-
-const INPUT_TOKEN = "WAVAX";
-const OUTPUT_TOKEN = "USDC";
-const DECIMAL_VALUE_IN = "0.001";
-const NATIVE_IN = true;
-const NATIVE_OUT = false;
-const SLIPPAGE = "0.5";
-
-async function main() {
-    await executeSwap(
-        INPUT_TOKEN,
-        OUTPUT_TOKEN,
-        DECIMAL_VALUE_IN,
-        NATIVE_IN,
-        NATIVE_OUT,
-        SLIPPAGE
-    );
-}
-
-main().catch((error) => {
-    console.error('Error in script:', error);
-});
